Adjust the package_dir to support develop mode
Adjust the package_dir argument in setup.py.
This way, when installing in development mode, the installed .pth file
will point to the src directory within the hapcat directory, rather than
the base of the hapcat repository.

This way, the module's actually able to be imported and everything
doesn't fail.